Fix player network latency (Player#getPing())

+++ src/main/java/cn/nukkit/Player.java
@@ -1081,9 +1081,10 @@ public class Player extends EntityHuman implements CommandSender, InventoryHolde
         }
         return 0;
     }
-
+    
     public int getPing() {
-        return this.interfaz.getNetworkLatency(this);
+        // TODO: Remove cast
+        return (int) this.interfaz.getNetworkLatency(this);
     }
 
     public boolean sleepOn(Vector3 pos) {
